Add types to more components
