+++ src/components/graphics/project-avatar-grid.module.css
